Added High Score System
Added a basic high score tracking system which resets each time the game is launched, but tracks between runs of the same instance. Updated text objects to reflect this, high score is shown in the upper right corner, and on the gameover screen. A New HIgh Score text is shown on gameover if the player got a new high score